# apps/api/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login with email/password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "tags": ["Authentication"]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "summary": "Refresh access token using refresh token",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Token refreshed successfully"
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        },
        "tags": ["Authentication"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "Logout (revoke refresh token)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logged out successfully"
          }
        },
        "tags": ["Authentication"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_registerUser",
        "summary": "Create a new user within the current organization",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          }
        },
        "tags": ["Authentication"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "summary": "Get current user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user retrieved"
          }
        },
        "tags": ["Authentication"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "VAT Management SaaS API",
    "description": "Multi-tenant VAT Management and IRD Submission API (Supabase PostgreSQL + RLS + White-labeling)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@abcvat.com"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": ["email", "password"]
      },
      "RegisterUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "member@abcvat.com"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "full_name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "role": {
            "type": "string",
            "enum": ["SUPER_ADMIN", "VAT_TEAM_LEAD", "VAT_TEAM_MEMBER", "CLIENT"],
            "default": "VAT_TEAM_MEMBER"
          }
        },
        "required": ["email", "password"]
      },
...
    }
  }
}